+++ frb_codegen/src/library/codegen/parser/file_reader.rs
@@ -1,0 +1,42 @@
+use std::path::{Path, PathBuf};
+use syn::File;
+use crate::codegen::dumper::Dumper;
+use crate::codegen::misc::GeneratorProgressBarPack;
+use crate::codegen::parser::reader::CachedRustReader;
+
+pub(crate) struct FileData {
+    pub(crate) path: PathBuf,
+    pub(crate) content: String,
+    pub(crate) ast: File,
+}
+
+pub(crate) fn read_files(
+    rust_input_paths: &[PathBuf],
+    rust_crate_dir: &Path,
+    cached_rust_reader: &mut CachedRustReader,
+    dumper: &Dumper,
+    progress_bar_pack: &GeneratorProgressBarPack,
+) -> anyhow::Result<Vec<FileData>> {
+    let _pb = progress_bar_pack.parse_cargo_expand.start();
+    let contents = rust_input_paths
+        .iter()
+        .map(|rust_input_path| {
+            let content =
+                cached_rust_reader.read_rust_file(rust_input_path, rust_crate_dir, dumper)?;
+            Ok((rust_input_path.to_owned(), content))
+        })
+        .collect::<anyhow::Result<Vec<(PathBuf, String)>>>()?;
+
+    contents
+        .into_iter()
+        .map(|(rust_input_path, content)| {
+            let ast = syn::parse_file(&content)?;
+            Ok(FileData {
+                path: rust_input_path,
+                content,
+                ast,
+            })
+        })
+        .collect()
+}
+
